app/requirements.txt
PyQt6
reportlab
openpyxl
python-barcode
Pillow 